fix: watch resources by id instead on omnictl delete
As ACLs can allow a user to have access to a single cluster, `omnictl delete cluster foo` fails even if the user have full access to the cluster `foo`. Actual deletion happens, but the attempt to watch all clusters and report when the deletion is complete fails, as the user does not have access to list/watch all clusters.

Instead, we watch the resources by their IDs, so the ACL engine allows the operation.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>