fix: implement `OngoingTasks` controller
Old implementation has a flaw: upgrade statuses don't have the right
creation time, as they are created with the cluster and stay idle while
there are no upgrades.

That's why we need a controller which will create and destroy ongoing
task resources.

This way the time on the resource will always reflect the time of
upgrade start.

Additionally fix `oneof` marshalling. Using standard `json.Marshal` does
not encode it correctly (there are no json tags on oneof), so use `protojson`
when the encoded spec is `proto.Message`.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>