feat: allow canceling deletion of a machine set node
This should help in case if the machine set node is destroyed, but the
machine set controller didn't scale it down yet (`ClusterMachine`
resource is not touched).
In that case it will be possible to restore the machine without
consequences.

Also did a slight refactoring of the nodes context menus, make a common
component for all node related context menus.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>