feat: add support for additional worker machine sets
Implement support for additional worker machine sets (creating them is only possible using cluster templates for the time being).

Fix the frontend so that additional worker sets are displayed and their config patches can be managed correctly.

Introduce the "name" concept for the worker machine sets. The name for the default machine set is `workers`, which preserves the current behavior, and follows the naming of `<cluster-id>-workers`. Additional sets follow the naming of `<cluster-id>-<name>`.

Closes siderolabs/omni#1116.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>