feat: implement `MachineClasses`
Fixes: https://github.com/siderolabs/omni/issues/1123

This PR adds a new resource called `MachineClass`.
`MachineClass` contains the list of machine label queries, which can be
used to filter the machines basing on some set of selectors.

The new page shows the list of machine classes and there is an
additional page for creating and editing machine classes.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>