feat: implement etcd backup controller
This PR implements the controller in charge of creating a backups (disabled for this PR). The controller uses
`EtcdBackupStore` interface to encrypt, upload and store backups. It also adds a validation to the new
`backupConfiguration` field in cluster spec.

It also provides a `EtcdBackupFileStore` as an implementation for `EtcdBackupStore` for storing backups locally in files.

Closes #1173
Closes #1176

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>