fix: make cluster status output more stable
Status is redrawn on each update, so for fluent output we need to make
sure items don't "jump" in the tree on each update, so we need some
stable sort order between the items, otherwise output is a mess.

The previous sort function was hard to read, and it was buggy leading to
random order between items.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>