feat: allow defining SAML label mapping rules to Omni roles
Introduce the new resource, `SAMLLabelRule` that contains label selectors and a role to be assigned to the user if the labels are matched on the `Identity` resources.

This makes it possible to assign SAML users to other roles than `None`.

Also, move the logic of syncing SAML labels with the Identity labels from ConfirmPublicKey step into the SAML session provider. which simplifies logic.

Closes siderolabs/omni#1271.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>