chore: add controller for `EtcdBackupOverallStatus` resource
We want to `EtcdBackupOverallStatus` to contain statuses about both etcd backup store and latest backup.
This commit changes store factories to put info into `EtcdBackupStoreStatus` and new `EtcdBackupOverallStatusController`
is now in charge of making `EtcdBackupOverallStatus`.

Also - update `EtcdBackupStatus` even if we failed to prepare the list of clusters to back up.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>