fix: prevent crashes in bootstrap manifest sync
Fixes #1394
Fixes #1395

Ignore "fatal" errors in the manifest sync which we know will not be
resolved by retrying the operation.

In the controller, save the fatal error instead of crashing.
In the UI, show '!' if the sync check had a fatal failure.

In the UI, fix handling of streaming API errors, and show the error in
the list of sync messages, so that a user knows why the sync failed.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>