fix: properly copy machine initial labels from the META partition
Machine status controller had a bug that was leading to removing all
user labels copied from the meta partition.
It got lost in the history of refactoring somehow.

Restructured machine status resource a bit: now initial labels are
always copied to the spec of the resource.
When deleting all labels and copying them from the `MachineLabel`
resource first copy labels from `spec.InitialLabels`.

This way it won't be possible to remove predefined labels, but it will
be possible to override it's value.

Also add the tests for it to avoid breaking this functionality in the
future.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>