feat: implement creating cluster new from an etcd snapshot
Allow creating a new cluster from an etcd snapshot.
This feature only works if the original cluster, i.e. the cluster that the backup is from, still exists (was not deleted).
In other words, this feature does not aim to bring back a deleted cluster, but provide a way to go back in time and inspect the past cluster state (point-in-time recovery).

Additionally, allow specifying the bootstrap spec in the cluster templates.

Closes siderolabs/omni#1378.
Closes siderolabs/omni#1376.
Closes siderolabs/omni#1430.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>