feat: implement exporting of cluster templates
Allow exporting a cluster template from an existing cluster.

It attempts to have minimal discrepancies when it is synced back (only updates on some config patches, and no deletion/creation).

This feature allows users to switch from managing a cluster via the UI to managing them by templates.

Also remove some unused code on the frontend and do small changes on frontend to reduce the discrepancy between the patches creates via the UI and via templates.

Closes siderolabs/omni#1206.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>