feat: replace docker imager with the image factory
Make Omni proxy requests to the image factory.
Update download code in the CLI and in the UI to first generate the
schematic and then call the download API.

Drop imager API, replace it with factory API, which is very similar, but
it proxies the requests to image factory instead of calling imager using
docker container run.

Change the installation media resource IDs to no longer be specific to
Omni, but use the same naming schema which is used in the image factory.

Omni specific names is now stored in spec. It also depends on Talos
version, so the final filename is composed on the fly in CLI and UI.

Add support for adding extensions, extra kernel args and using different
Talos versions in the install media loaders.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>