fix: remove basic auth functionality from the client
Basic auth is no more supported and was already removed from the server.

When configured on the client, it was causing the requests to have wrong auth headers and/or be not correctly signed.

Remove it from the client. When it is specified in omniconfig, issue a warning.

Remove the remaining traces of basic auth in the codebase.

Remove the AuthEnabledFunc from the signature interceptors, as Omni wouldn't work without auth and the check defaults to true when not set.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>