test: use machine UUID-based hostnames
Depends on https://github.com/siderolabs/talos/pull/8278
(need next Talos 1.7.0-alpha release to pull in talosctl).

The machines out of cluster are named `machine-$UUID`, once the join the
are named like `testname-cp-$UUID`. The idea is to make UUIDs searchable
in any logs we produce, and have meaningful hostnames when the machine
is in the cluster.

Note: this is only integration tests.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>