fix: template sync being "stuck"
When we added a bunch of CleanupControllers and other stuff which does
"magic" deletion of resources, the sync code was not updated to handle
it properly.

We can't assume that we're the only code that can destroy resources, so
handle destroy/teardown sequence in a way that it works even if
something else deletes the resources behind our back.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>